feat(git): Use AMR mode for diff-filter option
Currently, the diff-filter option in the `git diff command is AR. This
means that only added or renamed files are taken into account.

However, in a case where a migration is updated, we want to be able to
lint it anyways. So, changing the diff-filter mode to AMR should take
that into account.